+++ branches/restructure/edu.rice.cs.hpc.traceviewer/src/edu/rice/cs/hpc/traceviewer/painter/SpaceTimeMiniCanvas.java
@@ -164,10 +164,10 @@ public class SpaceTimeMiniCanvas extends SpaceTimeCanvas implements MouseListene
 		Point miniBottomRight = new Point((int)topLeftPixelX+viewingWidth, (int)topLeftPixelY+viewingHeight);
 		
 		long detailTopLeftTime = (long)(miniTopLeft.x/getScaleX());
-		double detailTopLeftProcess = miniTopLeft.y/getScaleY();
+		int detailTopLeftProcess = (int) (miniTopLeft.y/getScaleY());
 		
 		long detailBottomRightTime = (long)((double)miniBottomRight.x / getScaleX());
-		double detailBottomRightProcess = miniBottomRight.y/getScaleY();
+		int detailBottomRightProcess = (int) (miniBottomRight.y/getScaleY());
 		
 		detailCanvas.pushUndo();
 		detailCanvas.setDetailZoom(detailTopLeftTime, detailTopLeftProcess, detailBottomRightTime, detailBottomRightProcess);
